refactor(healthreporter): move package from pkg/system-probe to pkg/util
The healthreporter gRPC client is not specific to system-probe — it only
depends on the IPC component and the AgentSecure proto, both of which are
available to all sub-agents (process-agent, security-agent, etc.).

Move it to pkg/util/healthreporter so other sub-agents can reuse it without
taking a dependency on the system-probe package tree.
